move some fields from customization to defaults and css fixing
