+++ modules/web/src/app/core/components/sidenav/template.html
@@ -76,29 +76,6 @@ limitations under the License.
     </ng-container>
   </km-side-nav-field>
 
-    </ng-container>
-
-    <ng-container *ngIf="isSidenavCollapsed()">
-
-
-      <a mat-menu-item
-         fxLayoutAlign=" center"
-         [routerLink]="getRouterLink(view.Clusters)"
-         class="collapsed-sub-menu-item"
-         [ngClass]="{'active': checkUrl(view.Clusters), 'km-disabled': !getMenuItemClass(view.Clusters)}">
-        <span>Clusters</span>
-      </a>
-
-      <a mat-menu-item
-         *ngIf="areExternalClustersEnabled"
-         fxLayoutAlign=" center"
-         [routerLink]="getRouterLink(view.ExternalClusters)"
-         class="collapsed-sub-menu-item"
-         [ngClass]="{'active': checkUrl(view.ExternalClusters), 'km-disabled': !isDisabledField(view.ExternalClusters)}">
-        <span>External Clusters</span>
-      </a>
-    </ng-container>
-  </km-side-nav-field>
 
   <div [matTooltip]="getTooltip(view.ClusterTemplates)"
        [ngClass]="!isDisabledField(view.Overview) && 'km-disabled'"
