+++ modules/web/src/app/core/components/sidenav/theme.scss
@@ -88,46 +88,4 @@
       }
     }
   }
-
-  .sub-menu-item {
-    &:hover {
-      background-color: map.get($colors, sidenav-background-hover);
-    }
-
-    &.active {
-      a {
-        span {
-          color: map.get($colors, primary-dark);
-        }
-
-        &::after {
-          background-color: map.get($colors, primary);
-        }
-      }
-    }
-  }
-
-  .collapsed-sub-menu-item {
-    span {
-      color: map.get($colors, text);
-    }
-
-    &:hover {
-      background-color: map.get($colors, sidenav-background-hover) !important;
-
-      span {
-        color: map.get($colors, primary-dark);
-      }
-    }
-
-    &.active {
-      span {
-        color: map.get($colors, primary-dark);
-      }
-
-      &::after {
-        background-color: map.get($colors, primary);
-      }
-    }
-  }
 }
