+++ modules/web/src/app/settings/admin/customization/template.html
@@ -18,7 +18,7 @@ limitations under the License.
     <mat-card-title fxFlex
                     fxLayout="row"
                     fxLayoutAlign=" center">
-      <div fxFlex>Interface</div>
+      <div fxFlex>Customization</div>
     </mat-card-title>
   </mat-card-header>
   <mat-card-content>
@@ -57,86 +57,7 @@ limitations under the License.
           </div>
         </div>
       </div>
-<!-- 
-      <div fxLayout="row">
-        <div fxFlex="16%"
-             fxLayoutAlign=" center"
-             class="entry-label">
-          <span>Enable External Clusters</span>
-          <div class="km-icon-info km-pointer"
-               matTooltip="Enables external clusters in the UI and allows access to them through the API."></div>
-        </div>
-        <mat-checkbox [(ngModel)]="settings.enableExternalClusterImport"
-                      (change)="onSettingsChange()"
-                      id="km-enable-external-clusters-setting"></mat-checkbox>
-        <km-spinner-with-confirmation [isSaved]="isEqual(settings.enableExternalClusterImport, apiSettings.enableExternalClusterImport)"></km-spinner-with-confirmation>
-      </div>
 
-      <div fxLayout="row">
-        <div fxFlex="16%"
-             fxLayoutAlign=" center"
-             class="entry-label">
-          <span>Enable Kubernetes Dashboard</span>
-          <div class="km-icon-info km-pointer"
-               matTooltip="Show/Hide &quot;Open Dashboard&quot; button on cluster details and allow/block Kubernetes Dashboard access through the API."></div>
-        </div>
-        <mat-checkbox [(ngModel)]="settings.enableDashboard"
-                      (change)="onSettingsChange()"
-                      [disabled]="!isKubernetesDashboardFeatureGatesEnabled()"
-                      id="km-enable-kubernetes-dashboard-setting">
-          <ng-container *ngIf="!isKubernetesDashboardFeatureGatesEnabled()">
-            <mat-hint>This feature is disabled. Visit the
-              <a href="https://docs.kubermatic.com/kubermatic/{{editionVersion}}/tutorials-howtos/oidc-provider-configuration/share-clusters-via-delegated-oidc-authentication/"
-                 target="_blank"
-                 rel="noopener noreferrer">
-                documentation
-              </a>
-              to learn more.
-            </mat-hint>
-          </ng-container>
-        </mat-checkbox>
-        <km-spinner-with-confirmation [isSaved]="isEqual(settings.enableDashboard, apiSettings.enableDashboard)"></km-spinner-with-confirmation>
-      </div>
-
-      <div fxLayout="row">
-        <div fxFlex="16%"
-             fxLayoutAlign=" center"
-             class="entry-label">
-          <span>Enable OIDC Kubeconfig</span>
-          <div class="km-icon-info km-pointer"
-               matTooltip="Use OIDC provider as a proxy for kubeconfig download."></div>
-        </div>
-        <mat-checkbox [(ngModel)]="settings.enableOIDCKubeconfig"
-                      [disabled]="!isOIDCKubeCfgEndpointEnabled"
-                      (change)="onOIDCKubeconfigSettingsChange()"
-                      id="km-enable-oidc-setting">
-          <mat-hint *ngIf="!isOIDCKubeCfgEndpointEnabled">This feature is disabled. Visit the
-            <a href="https://docs.kubermatic.com/kubermatic/{{editionVersion}}/tutorials-howtos/oidc-provider-configuration/share-clusters-via-delegated-oidc-authentication/"
-               target="_blank"
-               rel="noopener noreferrer">
-              documentation
-            </a>
-            to learn more.
-          </mat-hint>
-        </mat-checkbox>
-        <km-spinner-with-confirmation [isSaved]="isEqual(settings.enableOIDCKubeconfig, apiSettings.enableOIDCKubeconfig)"></km-spinner-with-confirmation>
-      </div> -->
-
-
-      <div fxLayout="row">
-        <div fxFlex="16%"
-             fxLayoutAlign=" center"
-             class="entry-label">
-          <span>Enable WebTerminal</span>
-          <div class="km-icon-info km-pointer"
-               matTooltip="Show/Hide &quot;Web Terminal&quot; button on cluster details and allow/block web terminal pod access thorough terminal."></div>
-        </div>
-        <mat-checkbox [(ngModel)]="settings.enableWebTerminal"
-                      [disabled]="!settings.enableOIDCKubeconfig"
-                      (change)="onSettingsChange()">
-        </mat-checkbox>
-        <km-spinner-with-confirmation [isSaved]="isEqual(settings.enableWebTerminal, apiSettings.enableWebTerminal)"></km-spinner-with-confirmation>
-      </div>
 
     </div>
   </mat-card-content>
