+++ modules/web/src/app/settings/admin/defaults/template.html
@@ -121,7 +121,6 @@ limitations under the License.
           </div>
         </div>
 
-<!-- 
       <div fxLayout="row">
         <div fxFlex="16%"
              fxLayoutAlign=" center"
@@ -184,7 +183,23 @@ limitations under the License.
           </mat-hint>
         </mat-checkbox>
         <km-spinner-with-confirmation [isSaved]="isEqual(settings.enableOIDCKubeconfig, apiSettings.enableOIDCKubeconfig)"></km-spinner-with-confirmation>
-      </div> -->
+      </div>
+
+      <div fxLayout="row">
+        <div fxFlex="16%"
+             fxLayoutAlign=" center"
+             class="entry-label">
+          <span>Enable WebTerminal</span>
+          <div class="km-icon-info km-pointer"
+               matTooltip="Show/Hide &quot;Web Terminal&quot; button on cluster details and allow/block web terminal pod access thorough terminal."></div>
+        </div>
+        <mat-checkbox [(ngModel)]="settings.enableWebTerminal"
+                      [disabled]="!settings.enableOIDCKubeconfig"
+                      (change)="onSettingsChange()">
+        </mat-checkbox>
+        <km-spinner-with-confirmation [isSaved]="isEqual(settings.enableWebTerminal, apiSettings.enableWebTerminal)"></km-spinner-with-confirmation>
+      </div>
+
 
         <div fxLayout="row">
           <div fxFlex="16%"
