+++ modules/web/src/app/settings/admin/nav/style.scss
@@ -23,7 +23,7 @@ $sidenav-width-collapsed: 54px;
   @include mixins.size(variables.$sidenav-width, 100%);
 
   overflow-y: auto;
-  padding: (variables.$content-padding + variables.$toolbar-height) 0 0 10px;
+  padding-top: (variables.$content-padding + variables.$toolbar-height);
 
   .menu-item {
     height: $sidenav-item-height;
@@ -50,17 +50,61 @@ $sidenav-width-collapsed: 54px;
   }
 }
 
-.sub-menu-item {
-  height: 25px;
-  margin: 0 0 10px 45px;
-  padding: 5px 10px;
-}
-
 .collapse-sidenav {
   overflow-x: hidden;
   width: $sidenav-width-collapsed;
 
   .menu-item a i {
     margin-right: 0;
+  }
+}
+
+.sub-menu-item {
+  position: relative;
+
+  a {
+    margin-left: 22px;
+    padding: 6px 5px 6px 30px;
+
+    &:hover {
+      border-radius: variables.$border-radius;
+    }
+  }
+
+  &.active > a::after {
+    content: '';
+    height: 60%;
+    left: 22px;
+    position: absolute;
+    top: 20%;
+    width: 3px;
+  }
+}
+
+km-side-nav-field .sub-menu-item:last-child {
+  margin-bottom: 20px;
+}
+
+.menu-title {
+  border-bottom: 2px solid gray;
+}
+
+.collapsed-sub-menu-item {
+  height: 30px;
+  margin: 0 10px 0 15px;
+  padding: 5px 0 5px 25px;
+  position: relative;
+
+  &:hover {
+    border-radius: variables.$border-radius;
+  }
+
+  &.active::after {
+    content: '';
+    height: 60%;
+    left: 0;
+    position: absolute;
+    top: 20%;
+    width: 3px;
   }
 }
