+++ modules/web/src/app/settings/admin/nav/template.html
@@ -50,6 +50,31 @@ limitations under the License.
 
     </ng-container>
 
+    <ng-container *ngIf="isSidenavCollapsed">
+      <a mat-menu-item
+         fxLayoutAlign=" center"
+         routerLinkActive="active"
+         [routerLink]="getRouterLink(adminPanelView.Defaults)"
+         class="collapsed-sub-menu-item">
+        <span>{{adminPanelViewDisplayName.Defaults}}</span>
+      </a>
+
+      <a mat-menu-item
+      fxLayoutAlign=" center"
+      routerLinkActive="active"
+      [routerLink]="getRouterLink(adminPanelView.Limits)"
+      class="collapsed-sub-menu-item">
+     <span>{{adminPanelViewDisplayName.Limits}}</span>
+   </a>
+
+   <a mat-menu-item
+   fxLayoutAlign=" center"
+   routerLinkActive="active"
+   [routerLink]="getRouterLink(adminPanelView.Customization)"
+   class="collapsed-sub-menu-item">
+  <span>{{adminPanelViewDisplayName.Customization}}</span>
+</a>
+    </ng-container>
   </km-side-nav-field>
 
     <km-side-nav-field [label]="adminPanelMainSections.ManageResources"
@@ -112,8 +137,58 @@ limitations under the License.
           </a>
         </div>        
       </ng-container>
-    </km-side-nav-field>
 
+      <ng-container *ngIf="isSidenavCollapsed">
+        <a mat-menu-item
+        fxLayoutAlign=" center"
+        routerLinkActive="active"
+        [routerLink]="getRouterLink(adminPanelView.SeedConfiguration)"
+        class="collapsed-sub-menu-item">
+       <span>{{adminPanelViewDisplayName.SeedConfiguration}}</span>
+     </a>
+
+     <a mat-menu-item
+     fxLayoutAlign=" center"
+     routerLinkActive="active"
+     [routerLink]="getRouterLink(adminPanelView.Datacenters)"
+     class="collapsed-sub-menu-item">
+    <span>{{adminPanelViewDisplayName.Datacenters}}</span>
+  </a>
+
+  <a mat-menu-item
+  fxLayoutAlign=" center"
+  routerLinkActive="active"
+  [routerLink]="getRouterLink(adminPanelView.ProviderPresets)"
+  class="collapsed-sub-menu-item">
+ <span>{{adminPanelViewDisplayName.ProviderPresets}}</span>
+</a>
+
+<a mat-menu-item
+fxLayoutAlign=" center"
+routerLinkActive="active"
+[routerLink]="getRouterLink(adminPanelView.BackupDestinations)"
+class="collapsed-sub-menu-item">
+<span>{{adminPanelViewDisplayName.BackupDestinations}}</span>
+</a>
+
+<a mat-menu-item
+fxLayoutAlign=" center"
+routerLinkActive="active"
+[routerLink]="getRouterLink(adminPanelView.ProjectQuotas)"
+class="collapsed-sub-menu-item">
+<span>{{adminPanelViewDisplayName.ProjectQuotas}}</span>
+</a>
+
+<a mat-menu-item
+fxLayoutAlign=" center"
+routerLinkActive="active"
+[routerLink]="getRouterLink(adminPanelView.OPA)"
+class="collapsed-sub-menu-item">
+<span>{{adminPanelViewDisplayName.OPA}}</span>
+</a>
+      </ng-container>
+    </km-side-nav-field>
+    
     <km-side-nav-field [label]="adminPanelMainSections.Monitoring"
                      icon="km-icon-metering"
                      [isSidenavCollapsed]="isSidenavCollapsed">
@@ -137,6 +212,23 @@ limitations under the License.
           </a>
         </div>
       </ng-container>
+
+      <ng-container *ngIf="isSidenavCollapsed">
+        <a mat-menu-item
+        fxLayoutAlign=" center"
+        routerLinkActive="active"
+        [routerLink]="getRouterLink(adminPanelView.RuleGroups)"
+        class="collapsed-sub-menu-item">
+       <span>{{adminPanelViewDisplayName.RuleGroups}}</span>
+     </a>
+     <a mat-menu-item
+     fxLayoutAlign=" center"
+     routerLinkActive="active"
+     [routerLink]="getRouterLink(adminPanelView.Metering)"
+     class="collapsed-sub-menu-item">
+    <span>{{adminPanelViewDisplayName.Metering}}</span>
+  </a>
+</ng-container>
     </km-side-nav-field>
 
     <km-side-nav-field [label]="adminPanelMainSections.Users"
@@ -161,6 +253,22 @@ limitations under the License.
           </a>
         </div>
       </ng-container>
+      <ng-container *ngIf="isSidenavCollapsed">
+        <a mat-menu-item
+        fxLayoutAlign=" center"
+        routerLinkActive="active"
+        [routerLink]="getRouterLink(adminPanelView.Accounts)"
+        class="collapsed-sub-menu-item">
+       <span>{{adminPanelViewDisplayName.Accounts}}</span>
+     </a>
+     <a mat-menu-item
+     fxLayoutAlign=" center"
+     routerLinkActive="active"
+     [routerLink]="getRouterLink(adminPanelView.Administrators)"
+     class="collapsed-sub-menu-item">
+    <span>{{adminPanelViewDisplayName.Administrators}}</span>
+  </a>
+    </ng-container>
     </km-side-nav-field>
   <div fxFlex></div>
   <div class="menu-item km-pointer">
