+++ modules/web/src/app/settings/admin/nav/theme.scss
@@ -34,17 +34,49 @@
   }
 
   .sub-menu-item {
-    &:hover {
-      background-color: map.get($colors, sidenav-background-hover);
+    a {
+      &:hover {
+        background-color: map.get($colors, sidenav-background-hover);
+
+        span {
+          color: map.get($colors, primary-dark);
+        }
+      }
     }
 
     &.active {
-      border-left: 3px solid map.get($colors, primary);
-
       a {
         span {
           color: map.get($colors, primary-dark);
         }
+
+        &::after {
+          background-color: map.get($colors, primary);
+        }
+      }
+    }
+  }
+
+  .collapsed-sub-menu-item {
+    span {
+      color: map.get($colors, text);
+    }
+
+    &:hover {
+      background-color: map.get($colors, sidenav-background-hover) !important;
+
+      span {
+        color: map.get($colors, primary-dark);
+      }
+    }
+
+    &.active {
+      span {
+        color: map.get($colors, primary-dark);
+      }
+
+      &::after {
+        background-color: map.get($colors, primary);
       }
     }
   }
